[ADD-UNIT-TEST]: Add integration tests for REST endpoints.
With this commit we make a step toward integration testing with testcontainers. Testing the REST endpoints is not the final intent, but it is a start to begin work on the migration to (synchronous) gRPC services.

Update TODO list in the README.md to keep track of some missing features in this project.