[ADD-UNIT-TEST]: Add unit test classes and make use of game design data for reward calculation.
Add unit test base class and start implementing unit test for all the necessary classes. Rename all the test classes wherever needed in order to state what type of test it is being implemented. Introduce the first version of game design data to perform the calculation of rewards on box occupied and on stage cleared.